Revert "Rename test case and fix code style issues"
This reverts commit 43a88c87d6f9b550205f14ff793e9815ba99ff9a.
