+++ src/core/etl/tests/Flow/ETL/Tests/Unit/Extractor/ExtractorTestCase.php
@@ -2,32 +2,13 @@
 
 declare(strict_types=1);
 
-namespace Flow\ETL\Tests;
+namespace Flow\ETL\Tests\Unit\Extractor;
 
 use Flow\ETL\{Config, Extractor, FlowContext, Rows};
 use PHPUnit\Framework\TestCase;
 
-abstract class FlowTestCase extends TestCase
+abstract class ExtractorTestCase extends TestCase
 {
-    public function assertExtractorCountBatches(int $expectedCount, Extractor $extractor) : void
-    {
-        static::assertCount(
-            $expectedCount,
-            \iterator_to_array($extractor->extract(new FlowContext(Config::default())))
-        );
-    }
-
-    public function assertExtractorCountRows(int $expectedCount, Extractor $extractor) : void
-    {
-        $totalRows = 0;
-
-        foreach ($extractor->extract(new FlowContext(Config::default())) as $rows) {
-            $totalRows += $rows->count();
-        }
-
-        static::assertSame($expectedCount, $totalRows);
-    }
-
     public function assertExtractorCountRowsPerBatch(int $expectedCount, Extractor $extractor) : void
     {
         $extractorContainsBatches = false;
@@ -42,14 +23,21 @@ abstract class FlowTestCase extends TestCase
         }
     }
 
-    /**
-     * @param array<Rows> $expectedRows
-     * @param Extractor $extractor
-     */
-    public function assertExtractorEqualsRows(array $expectedRows, Extractor $extractor) : void
+    public function assertExtractorCountRows(int $expectedCount, Extractor $extractor) : void
     {
-        static::assertEquals(
-            $expectedRows,
+        $totalRows = 0;
+
+        foreach ($extractor->extract(new FlowContext(Config::default())) as $rows) {
+            $totalRows += $rows->count();
+        }
+
+        static::assertSame($expectedCount, $totalRows);
+    }
+
+    public function assertExtractorCountBatches(int $expectedCount, Extractor $extractor) : void
+    {
+        static::assertCount(
+            $expectedCount,
             \iterator_to_array($extractor->extract(new FlowContext(Config::default())))
         );
     }
@@ -67,5 +55,17 @@ abstract class FlowTestCase extends TestCase
         }
 
         static::assertSame($expectedArray, $data);
+    }
+
+    /**
+     * @param array<Rows> $expectedRows
+     * @param Extractor $extractor
+     */
+    public function assertExtractorEqualsRows(array $expectedRows, Extractor $extractor) : void
+    {
+        static::assertEquals(
+            $expectedRows,
+            \iterator_to_array($extractor->extract(new FlowContext(Config::default())))
+        );
     }
 }
