+++ src/core/etl/tests/Flow/ETL/Tests/Unit/Extractor/SequenceExtractorTest.php
@@ -9,9 +9,9 @@ use function Flow\ETL\DSL\{date_entry,
     from_sequence_date_period,
     from_sequence_date_period_recurrences,
     from_sequence_number};
-use Flow\ETL\{Row, Rows, Tests\FlowTestCase};
+use Flow\ETL\{Row, Rows};
 
-final class SequenceFlowTest extends FlowTestCase
+final class SequenceExtractorTest extends ExtractorTestCase
 {
     public function test_extracting_from_date_period() : void
     {
